// src/modules/csv/csv.interface.ts
export interface ICSVFile {
	id: string;
	apoiador: string;
	cpfcnpj: string;
	email: string;
	telefone: string;
	valor: string;
	faixaderecompensa: string;
	competenciadoapoio: string;
	status: string;
	descricaodostatus: string;
	metododepagamento: string;
	cep: string;
	rua: string;
	numero: string;
	complementos: string;
	bairro: string;
	cidade: string;
	uf: string;
	pais: string;
	enderecocompleto: string;
}